wordpress: Make Database and Table into public functions.
This was useful when constructing a manual sqrl.Select query to generate index of sub-sitemaps compatible with a plugin currently used on a personal blog, and which needs to contain a somewhat accurate 'last modified' value. The constructed sqrl query needs access to 'posts' table, and invocation using q.Run requires the database itself, so making this public allows both to be 'extracted' from the context object and allows for a custom query to be submitted without having to establish another custom connection for either sqlr or something else to use.